Extracted Shedpi module utils
