--- shedpi_module_utils/pyproject.toml
@@ -1,14 +1,0 @@
-[tool.poetry]
-name = "shedpi-module-utils"
-version = "0.1.0"
-description = ""
-authors = ["Andrew Aikman <aiky30@yahoo.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.11"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
